requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests~=2.32.3
pydantic==2.4.2
pydantic-settings==2.1.0

python-dotenv==1.0.0
PyGithub==2.1.1
pytest==7.4.3
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1